+++ R/bmm-package.R
@@ -2,5 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
+#' @importFrom brms save_pars
 ## usethis namespace: end
 NULL