+++ R/bmm_model_mixture3p.R
@@ -2,7 +2,8 @@
 # MODELS                                                                 ####
 #############################################################################!
 
-.model_mixture3p <- function(resp_err, nt_features, setsize, ...) {
+.model_mixture3p <- function(resp_err = NULL, nt_features = NULL, setsize = NULL, regex = FALSE, ...) {
+
   out <- list(
     resp_vars = nlist(resp_err),
     other_vars = nlist(nt_features, setsize),
@@ -30,24 +31,30 @@
       )),
     void_mu = FALSE
   )
+  attr(out, "regex") <- regex
+  attr(out, "regex_vars") <- c('nt_features') # variables that can be specified via regular expression
   class(out) = c("bmmmodel", "vwm", "nontargets", "mixture3p")
   out
 }
 
 
 # user facing alias
-#' @title `r .model_mixture3p(NA, NA, NA)$info$name`
-#' @details `r model_info(mixture3p(NA, NA, NA))`
+#' @title `r .model_mixture3p()$info$name`
+#' @details `r model_info(.model_mixture3p())`
 #' @param resp_err The name of the variable in the dataset containing
 #'   the response error. The response error should code the response relative to
 #'   the to-be-recalled target in radians. You can transform the response error
 #'   in degrees to radians using the `deg2rad` function.
 #' @param nt_features A character vector with the names of the non-target
-#'   feature values. The non_target feature values should be in radians and centered
-#'   relative to the target.
+#'   feature values. The non_target feature values should be in radians and
+#'   centered relative to the target. Alternatively, if regex=TRUE, a regular
+#'   expression can be used to match the non-target feature columns in the
+#'   dataset.
 #' @param setsize Name of the column containing the set size variable (if
 #'   setsize varies) or a numeric value for the setsize, if the setsize is
 #'   fixed.
+#' @param regex Logical. If TRUE, the `nt_features` argument is interpreted as
+#'  a regular expression to match the non-target feature columns in the dataset.
 #' @param ... used internally for testing, ignore it
 #' @return An object of class `bmmmodel`
 #' @keywords bmmmodel
@@ -69,18 +76,34 @@
 #'   thetant ~ 1
 #' )
 #'
-#' # specify the 3-parameter model
-#' model <- mixture3p(resp_err = "y", nt_features = paste0('nt',1:3,'_loc'), setsize = 4)
+#' # specify the 3-parameter model with explicit column names for non-target features
+#' model1 <- mixture3p(resp_err = "y", nt_features = paste0('nt',1:3,'_loc'), setsize = 4)
 #'
 #' # fit the model
 #' fit <- fit_model(formula = ff,
 #'                  data = dat,
-#'                  model = model,
+#'                  model = model1,
+#'                  parallel=T,
+#'                  iter = 500,
+#'                  backend='cmdstanr')
+#'
+#' # alternatively specify the 3-parameter model with a regular expression to match non-target features
+#' # this is equivalent to the previous call, but more concise
+#' model2 <- mixture3p(resp_err = "y", nt_features = "nt.*_loc", setsize = 4, regex = TRUE)
+#'
+#' # fit the model
+#' fit <- fit_model(formula = ff,
+#'                  data = dat,
+#'                  model = model2,
 #'                  parallel=T,
 #'                  iter = 500,
 #'                  backend='cmdstanr')
 #' }
-mixture3p <- .model_mixture3p
+mixture3p <- function(resp_err, nt_features, setsize, regex = FALSE, ...) {
+  stop_missing_args()
+  .model_mixture3p(resp_err = resp_err, nt_features = nt_features,
+                   setsize = setsize, regex = regex, ...)
+}
 
 #############################################################################!
 # CONFIGURE_MODEL METHODS                                                ####
@@ -130,15 +153,19 @@ configure_model.mixture3p <- function(model, data, formula) {
   additional_constants <- list()
   additional_constants[[kappa_unif]] <- -100
   additional_constants[[mu_unif]] <- 0
-  prior <- fixed_pars_priors(model, additional_constants) +
-    brms::prior_("normal(2, 1)", class = "b", nlpar = "kappa") +
-    brms::prior_("logistic(0, 1)", class = "b", nlpar = "thetat") +
-    brms::prior_("logistic(0, 1)", class = "b", nlpar = "thetant")
+  prior <- fixed_pars_priors(model, additional_constants)
+  if (getOption("bmm.default_priors", TRUE)) {
+    prior <- prior +
+      set_default_prior(bmm_formula, data,
+                        prior_list=list(kappa=list(main='normal(2,1)',effects='normal(0,1)', nlpar=T),
+                                        thetat=list(main='logistic(0, 1)', nlpar=T),
+                                        thetant=list(main='logistic(0, 1)', nlpar=T)))
+  }
 
   # if there is setsize 1 in the data, set constant prior over thetant for setsize1
-  if ((1 %in% data$ss_numeric) && !is.numeric(data[[setsize_var]])) {
-    prior <- prior +
-      brms::prior_("constant(-100)", class="b", coef = paste0(setsize_var, 1), nlpar="thetant")
+  thetant_preds <- rhs_vars(bmm_formula$thetant)
+  if (any(data$ss_numeric == 1) && !is.numeric(data[[setsize_var]]) && setsize_var %in% thetant_preds) {
+    prior <- combine_prior(prior, brms::prior_("constant(-100)", class="b", coef = paste0(setsize_var, 1), nlpar="thetant"))
   }
 
   nlist(formula, data, family, prior)