+++ R/fit_model.R
@@ -24,6 +24,12 @@
 #'   or related functions and combined using the c method or the + operator. See
 #'   also [get_model_prior()] for more help. Not necessary for the default model
 #'   fitting, but you can provide prior constraints to model parameters
+#' @param sort_data Logical. If TRUE, the data will be sorted by the predictor
+#'   variables for faster sampling. If FALSE, the data will not be sorted, but
+#'   sampling will be slower. If "check" (the default), `fit_model()` will check if
+#'   the data is sorted, and ask you via a console prompt if it should be
+#'   sorted. You can set the default value for this option using global
+#'   `options(bmm.sort_data = TRUE/FALSE/"check)`)` or via `bmm_options(sort_data)`
 #' @param silent Verbosity level between 0 and 2. If 1 (the default), most of the
 #'   informational messages of compiler and sampler are suppressed. If 2, even
 #'   more messages are suppressed. The actual sampling progress is still
@@ -68,24 +74,31 @@
 #'                  backend='cmdstanr')
 #' }
 #'
-fit_model <- function(formula, data, model, parallel = FALSE, chains = 4,
-                      prior = NULL, silent = getOption('bmm.silent', 1), ...) {
+fit_model <- function(formula, data, model,
+                      prior = NULL,
+                      chains = 4,
+                      parallel = getOption('bmm.parallel', FALSE),
+                      sort_data = getOption('bmm.sort_data', "check"),
+                      silent = getOption('bmm.silent', 1),
+                      ...) {
   # warning for using old version
   dots <- list(...)
   if ("model_type" %in% names(dots)) {
     stop('The "model_type" argument was deprecated on Feb 3, 2024. Either:
          - See ?fit_model for the new usage;
-         - or install the old version of the package with: devtools::install_github("venpopov/bmm@v0.0.1")
-         ')
+         - or install the old version of the package with: devtools::install_github("venpopov/bmm@v0.0.1")')
   }
 
   # set temporary global options and return modified arguments for brms
-  opts <- configure_options(nlist(parallel, chains, silent))
+  configure_opts <- nlist(parallel, chains, sort_data, silent)
+  opts <- configure_options(configure_opts)
 
   # check model, formula and data, and transform data if necessary
-  model <- check_model(model)
-  formula <- check_formula(model, formula)
+  model <- check_model(model, data)
   data <- check_data(model, data, formula)
+  user_formula <- formula
+  formula <- check_formula(model, data, formula)
+
 
   # generate the model specification to pass to brms later
   config_args <- configure_model(model, data, formula)
@@ -99,22 +112,7 @@ fit_model <- function(formula, data, model, parallel = FALSE, chains = 4,
   fit <- call_brm(fit_args)
 
   # model postprocessing
-  postprocess_brm(model, fit)
-}
-
-
-
-#' @export
-update.bmmfit <- function(fit, ...) {
-  stop("The update method for bmmfit is not yet implemented, but is planned for a future release.", call. = FALSE)
-  # do any necessary preprocessing to accomondate bmm changes into brms
-  # ....
-
-  fit = NextMethod(fit)  # pass back to brms::update.brmsfit
-
-  # do any necessary postprocessing to convert back to bmmfit
-  # ....
-  class(fit) <- c('bmmfit', class(fit))  # reassing class
-  return(fit)
+  postprocess_brm(model, fit, fit_args = fit_args, user_formula = user_formula,
+                  configure_opts = configure_opts)
 }