+++ R/helpers-prior.R
@@ -65,9 +65,9 @@ combine_prior <- function(prior1, prior2) {
 #' }
 #'
 get_model_prior <- function(formula, data, model, ...) {
-  model <- check_model(model)
-  formula <- check_formula(model, formula)
+  model <- check_model(model, data)
   data <- check_data(model, data, formula)
+  formula <- check_formula(model, data, formula)
   config_args <- configure_model(model, data, formula)
 
   dots <- list(...)
@@ -103,3 +103,133 @@ fixed_pars_priors <- function(model, additional_pars = list()) {
   brms::set_prior(priors, class = "Intercept", dpar = pars)
 }
 
+
+#' Set default priors for a bmmmodel
+#'
+#' For developers to use within configure_model.* functions. This function
+#' allows you to specify default priors flexibly regardless of the formula the
+#' user has supplied. The function will automatically recognize when intercepts
+#' are present or suppressed, and will set the default priors accordingly. You
+#' can specify priors of intercepts/main levels of a predictor with supressed
+#' intercept, and priors on the effects of the predictors relative to the
+#' intercept.
+#'
+#' @param bmmformula A `bmmformula` object
+#' @param data A data.frame containing the data used in the model
+#' @param prior_list A list of lists containing the priors for the parameters.
+#'   The list should have the same names as the parameters in the `bmmformula`
+#'   for which you want to set the default prior. Each parameter should be
+#'   assigned a list of priors. The 'main' entry will be used for the Intercept
+#'   or any of the main levels if the intercept is suppressed. The 'effects' entry
+#'   is optional. If given, it will be used as the prior for effects of the
+#'   predictor relative to the intercept. Finally, an 'nlpar = TRUE' entry can be given
+#'   in the list for each predictor, to specify that the prior should be set on
+#'   `nlpar` (see `bmm_model_mixture3p.R` for an example). By default the priors
+#'   will be set on dpar
+#' @examples
+#' data <- OberauerLin_2017
+#' data$session <- as.factor(data$session)
+#' # suppressed intercept on thetat, intercept present for kappa
+#' formula <- bmf(thetat ~ 0 + set_size, kappa ~ session)
+#' prior_list <- list(thetat = list(main = 'logistic(0,1)', nlpar=TRUE),
+#'                    kappa = list(main = 'normal(2,1)', effects = 'normal(0,1)', nlpar=TRUE))
+#' prior <- set_default_prior(formula, data, prior_list)
+#' print(prior)
+#'
+#' # suppressed intercept on both thetat and kappa
+#' formula <- bmf(thetat ~ 0 + set_size, kappa ~ 0 + session)
+#' prior_list <- list(thetat = list(main = 'logistic(0,1)', nlpar=TRUE),
+#'                    kappa = list(main = 'normal(2,1)', effects = 'normal(0,1)', nlpar=TRUE))
+#' prior <- set_default_prior(formula, data, prior_list)
+#' print(prior)
+#'
+#' # suppressed intercept on both thetat and kappa, with interaction for kappa
+#' formula <- bmf(thetat ~ 0 + set_size, kappa ~ 0 + set_size*session)
+#' prior_list <- list(thetat = list(main = 'logistic(0,1)', nlpar=TRUE),
+#'                    kappa = list(main = 'normal(2,1)', effects = 'normal(0,1)', nlpar=TRUE))
+#' prior <- set_default_prior(formula, data, prior_list)
+#' print(prior)
+#' @export
+#' @keywords internal, developer
+set_default_prior <- function(bmmformula, data, prior_list) {
+  dpars <- names(bmmformula)
+  pars_key <- names(prior_list)
+  prior <- brms::empty_prior()
+  if (any(not_in(pars_key, dpars))) {
+    stop("You are trying to set a default prior on a parameter that is not part of the model")
+  }
+  if (!is.list(prior_list)) {
+    stop("The prior_list should be a list of lists")
+  }
+  for (i in 1:length(prior_list)) {
+    if(!is.list(prior_list[[i]])) {
+      stop("The prior_list should be a list of lists")
+    }
+  }
+
+  pars <- dpars[dpars %in% pars_key]
+  is_nlpar <- sapply(prior_list[pars], function(x) {isTRUE(x$nlpar)})
+  for (par in pars) {
+    bform <- bmmformula[[par]]
+    bterms <- stats::terms(bform)
+    prior_desc <- prior_list[[par]]
+    has_effects_prior <- !is.null(prior_desc$effects)
+
+    all_rhs_names <- rhs_vars(bform)
+    all_rhs_terms <- attr(bterms, "term.labels")
+    fixef <- all_rhs_terms[all_rhs_terms %in% all_rhs_names]
+    inter <- all_rhs_terms[attr(bterms,'order') > 1]
+    nfixef <- length(fixef)
+    ninter <- length(inter)
+    interaction_only <- nfixef == 0 && ninter > 0
+
+    ## if the user has specified a non-linear predictor on a model parameter, do not set prior
+    if (any(all_rhs_names %in% dpars)) {
+      next
+    }
+
+    # # by default set the effects prior on the class 'b'. The intercept can be overwritten later
+    if (has_effects_prior && nfixef > 0) {
+      if (is_nlpar[par]) {
+        prior <- combine_prior(prior, brms::prior_(prior_desc$effects, class = "b", nlpar = par))
+      } else {
+        prior <- combine_prior(prior, brms::prior_(prior_desc$effects, class = "b", dpar = par))
+      }
+    }
+
+    # check if intercept is present and set prior_desc[[1]] on the intercept
+    if (attr(bterms, "intercept")) {
+      if (is_nlpar[par]) {
+        prior <- combine_prior(prior, brms::prior_(prior_desc$main, class = "b", coef = "Intercept", nlpar = par))
+      } else {
+        prior <- combine_prior(prior, brms::prior_(prior_desc$main, class = "Intercept", dpar = par))
+      }
+      next
+    }
+
+    # next check if there is only one predictor, in which case set the main prior on all levels
+    # same if there are multiple predictors, but they are specified only as an interaction
+    # get individual predictors, and the formula terms. Fixed effects are those that match
+    if ((nfixef == 1 && ninter == 0) || interaction_only) {
+      if (is_nlpar[par]) {
+        prior <- combine_prior(prior, brms::prior_(prior_desc[[1]], class = "b", nlpar = par))
+      } else {
+        prior <- combine_prior(prior, brms::prior_(prior_desc[[1]], class = "b", dpar = par))
+      }
+      next
+    }
+
+    # if there are multiple predictors, set the main prior on the levels of the first predictor
+    first_term <- attr(bterms,"term.labels")[1]
+    levels <- levels(data[[first_term]])
+    coefs <- paste0(first_term, levels)
+    for (coef in coefs) {
+      if (is_nlpar[par]) {
+        prior <- combine_prior(prior, brms::prior_(prior_desc[[1]], class = "b", coef = coef, nlpar = par))
+      } else {
+        prior <- combine_prior(prior, brms::prior_(prior_desc[[1]], class = "b", coef = coef, dpar = par))
+      }
+    }
+  }
+  prior
+}